Fixed a couple of tests to match recent changes in YAGL

+++ records/graphics/SpriteWrapperRecord.cpp
@@ -32,7 +32,7 @@ void SpriteWrapperRecord::write(std::ostream& os, const GRFInfo& info) const
     // sprite ID
     std::ostringstream ss;
     m_sprite->write(ss, info);
-    uint32_t size = ss.str().size();
+    uint32_t size = static_cast<uint32_t>(ss.str().size());
     
     write_uint32(os, m_sprite_id);
     write_uint32(os, size + 1); // +1 to account for the extra 0xFF.
